lambda/gpt-response/requirements.txt
boto3==1.26.108
langchain==0.0.133
openai==0.27.4
pinecone_client==2.2.1
python-dotenv==1.0.0
requests==2.28.2
tiktoken==0.3.3
